backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=3.1.1
python-telegram-bot>=20.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
aiohttp>=3.8.0 